long long error to be fixed

+++ src/learner.h
@@ -27,8 +27,8 @@ class learner {
   int evaluate (intersection *node);
   int evaluate (void);
   int evaluate (intersection **nodes, int intc);
-  float* get_q_entry(int* state, int action);
-  float* get_max_q_entry(int* next_state);
+  long double* get_q_entry(int* state, int action);
+  long double* get_max_q_entry(int* next_state);
   float get_reward();
 
   //vars
@@ -41,7 +41,7 @@ class learner {
   int* sense_state();
   int select_action(int* curr_state);
   void apply_action(int action);
-  float* q_table;
+  long double* q_table;
 
   int* state;
   int action;
